refactor(Graph): Extract GraphCenterPanel from CenterPanel for reusability
