+++ src/pages/Graph/components/CenterPanel.jsx
@@ -1,90 +1,23 @@
-import  { useEffect, useState } from "react";
-import { motion } from "framer-motion";
-import { MyConsoleContext } from "../../../context/ConsoleContext";
-import { useDataVisualization } from "../../../hooks/use-datavisualization";
-import useWindowSize from "../../../hooks/use-windowsize";
-import { mobilmenu } from "../../../locales/localdata";
-const CenterPanel = () => {
-  const {
-    animationVariants,
-    mobiletoggle,
-    setMobileToggle,
-    statusTable,
-    message,
-  } = MyConsoleContext();
-  const { consoleContent } = useDataVisualization();
-  const { width, height } = useWindowSize();
-  const [activeMenu, setActiveMenu] = useState("dontshow"); //Status for the selected menu item
-  const breakpoints = {
-    sm: 640,
-    md: 768,
-    lg: 1024,
-    xl: 1200,
-    "2xl": 1700,
-  };
+import { motion } from 'framer-motion';
+import { MyConsoleContext } from '../../../context/ConsoleContext'; // Ellenőrizd az elérési utat!
 
-  const menuHandler = (item) => {
-    if (item.title === "showmenu") {
-      setMobileToggle(true);
-    } else {
-      setMobileToggle(false);
-    }
-    setActiveMenu(item.title); // Update selected menu item
-  };
-  //glowy-button-8
-  const fullglass = "  ";
+// Ez a komponens most már csak a külső keret és az animáció.
+// A tényleges tartalmat (pl. GraphCenterContent, MobileMenu) children-ként kapja.
+const CenterPanel = ({ children }) => {
+  // Csak az animációhoz szükséges kontextusra van szüksége (ha van)
+  const { animationVariants } = MyConsoleContext();
+
   return (
     <motion.div
-      animate={animationVariants}
-      className=" flex flex-col items-center justify-center relative z-10 rounded inset-0
-        border-0 border-lime-400  text-dark w-screen md:w-full h-full  shadow-2xl  p-2 "
+      animate={animationVariants} // Az animáció marad, ha szükséges
+      className="flex flex-col items-center justify-center relative z-10 rounded inset-0 border-0 border-lime-400 text-dark w-screen md:w-full h-full shadow-2xl p-2"
+     
     >
-    <div
-        className={`${
-          statusTable.dashboard === "graph" ? fullglass : ""
-        }    glowy-button-10  rounded-2xl
-      z-50 relative p-0 overflow-y-scroll overflow-x-scroll 
-         text-white flex flex-col   items-center justify-center w-screen 
-         xl:w-full md:h-full`}
-        // style={{ width: width - 800, height: height - 120 }}
-        style={
-          width >= breakpoints["2xl"]
-            ? { width: width - 600, height: height - 150 } // 2XL gets a 150 adjustment
-            : width >= breakpoints.xl
-            ? { width: width - 600, height: height - 150 } // XL gets a 50 adjustment
-            : { width: width - 20, height: height - 250 } // Other sizes
-        }
-      >
-        {consoleContent}
-      </div>
-
-      <div className="absolut  xl:hidden z-50  border-0 border-orange-400 w-full flex justify-center items-center my-5">
-        <div className="flex  border-0 border-red-400">
-          {mobilmenu.map((item, index) => (
-            <div
-              key={index}
-              onClick={() => menuHandler(item)}
-              className={`${
-                activeMenu === item.title
-                  ? "glowy-button-6 hover-active border-red-400 border-0"
-                  : ""
-              }  group relative z-50 border-0 hover:border-red-400 p-2 rounded m-2 bg-950 text-primary cursor-pointer text-3xl hover:bg-600`}
-            >
-              {<item.icon />}
-              <div className="hidden text-center absolute top-0 left-0 mt-[80px] ml-2 w-32  group-hover:block bg-gray-700 text-white text-sm rounded p-1 z-50">
-                {item.description}
-              </div>
-            </div>
-          ))}
-        </div>
-      </div>
+      {children}
     </motion.div>
   );
 };
 
 export default CenterPanel;
-
-
-
 
 //CenterPanel
